updated filing structure
+simple navigation (nothing permanent)
