--- index.html
@@ -1,18 +1,0 @@
-<!doctype html>
-  <html>
-    <head>
-      <link rel="stylesheet" type="text/css" href="stylesheet.css"
-      <meta charset="utf-8">
-    </head>
-    <body>
-      <h1>Hello World!</h1>
-      <p>
-        We are using node <script>document.write(process.versions.node)</script>,
-        Chrome <script>document.write(process.versions.chrome)</script>,
-        and Electron <script>document.write(process.versions.electron)</script>.
-      </p>
-      <p>
-      You shouldn't be seeing this, contact the developer.
-      </p>
-    </body>
-  </html>
