+++ src/about.html
@@ -1,0 +1,23 @@
+<!doctype html>
+  <html>
+    <head>
+      <link rel="stylesheet" type="text/css" href="stylesheet.css"
+      <meta charset="utf-8">
+    </head>
+    <body>
+      <h1>About Siren</h1>
+      <center>
+        <ul>
+          <li><a href="index.html">Home</a></li><li><a href="about.html">About</a></li><li><a href="music.html">Music</a></li>
+        </ul>
+      </center>
+        <p>
+          Siren is using:
+        </p>
+        <p>
+          Node.js Version: <script>document.write(process.versions.node)</script>,
+          Chrome Version: <script>document.write(process.versions.chrome)</script>,
+          and Electron Version: <script>document.write(process.versions.electron)</script>.
+        </p>
+    </body>
+  </html>
