+++ src/index.html
@@ -1,0 +1,20 @@
+<!doctype html>
+  <html>
+    <head>
+      <link rel="stylesheet" type="text/css" href="stylesheet.css"
+      <meta charset="utf-8">
+    </head>
+    <body>
+      <h1>Welcome To Siren</h1>
+      <p>
+        <center>
+          <ul>
+            <li><a href="index.html">Home</a></li><li><a href="about.html">About</a></li><li><a href="music.html">Music</a></li>
+          </ul>
+        </center>
+      </p>
+      <p>
+        index
+      </p>
+    </body>
+  </html>
