+++ src/music.html
@@ -1,0 +1,19 @@
+<!doctype html>
+  <html>
+    <head>
+      <link rel="stylesheet" type="text/css" href="stylesheet.css"
+      <meta charset="utf-8">
+    </head>
+    <body>
+      <h1>About Siren</h1>
+      <br/>
+      <center>
+        <ul>
+          <li><a href="index.html">Home</a></li><li><a href="about.html">About</a></li><li><a href="music.html">Music</a></li>
+        </ul>
+      </center>
+        <p>
+          Music Goes Here!
+        </p>
+    </body>
+  </html>
