+++ src/stylesheet.css
@@ -1,0 +1,85 @@
+@import url('https://fonts.googleapis.com/css?family=Montserrat|Roboto&subset=greek');
+
+body {
+}
+
+h1 {
+  text-align: center;
+  font-family: 'Montserrat', sans-serif;
+}
+
+P {
+  font-family: verdana;
+  font-size: 15px;
+  line-height: normal;
+}
+
+ul {
+  padding: 0;
+  width: 100%;
+}
+
+li {
+  display: inline;
+}
+
+a {
+  outline: none;
+  text-decoration: none;
+  display: inline-block;
+  width: 19.5%;
+  margin-right: 0.625%;
+  text-align: center;
+  line-height: 3;
+  color: black;
+  font-family: 'Montserrat', sans-serif;
+}
+
+li:last-child a {
+  margin-right: 0;
+}
+
+a:link, a:visited, a:focus {
+  background: darkorange;
+}
+
+a:hover {
+  background: #717171;
+}
+
+a:active {
+  background: #717171;
+  color: white;
+}
+
+.splashheader {
+  font-family: 'Montserrat', sans-serif;
+}
+
+.splashbody {
+  text-align: center;
+  user-select: none;
+  -webkit-app-region: drag;
+  font-family: 'Roboto', sans-serif;
+}
+
+.container {
+  height: 10em;
+  position: relative;
+  letter-spacing: 1.5px;
+}
+
+.text {
+  margin: 0;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  margin-right: -50%;
+  transform: translate(-50%, -50%);
+  font-size: 20px;
+  outline-color: #ffffff;
+}
+
+.dud {
+  color: #717171;
+}
