--- stylesheet.css
@@ -1,44 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Montserrat|Roboto&subset=greek');
-
-body {
-}
-
-h1 {
-  text-align: center;
-}
-
-P {
-  font-family: verdana;
-  font-size: 15px;
-}
-
-.splashheader {
-  font-family: verdana, cursive;
-}
-
-.splashbody {
-  text-align: center;
-  user-select: none;
-  -webkit-app-region: drag;
-  font-family: 'Roboto', sans-serif;
-}
-
-.container {
-  height: 10em;
-  position: relative;
-  letter-spacing: 1.5px;
-}
-
-.text {
-  margin: 0;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  margin-right: -50%;
-  transform: translate(-50%, -50%);
-  font-size: 20px;
-}
-
-.dud {
-  color: #757575;
-}
